Move requestJetpackToken to its own utils function and refactor.
