+++ packages/js/ai/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './text-completion';
 export * from './create-extended-error';
+export * from './requestJetpackToken';
